Update ve insert çalışır hale getirildi, hatalar olabilir

+++ src/Employees.java
@@ -11,11 +11,11 @@ public class Employees {
 	Integer salary;
 	Integer manager_id;
 	Integer department_id;
-	Integer commission_pct;
+	Double commission_pct;
 
 
 	public Employees(Integer employee_id, String first_name, String last_name, String email, String phone_number, String hire_date,
-			 String job_id, Integer salary, Integer commission_pct, Integer manager_id, Integer department_id) {
+			 String job_id, Integer salary, Double commission_pct, Integer manager_id, Integer department_id) {
 		super();
 		this.employee_id = employee_id;
 		this.first_name = first_name;
@@ -106,11 +106,11 @@ public class Employees {
 		this.department_id = department_id;
 	}
 
-	public Integer getCommission_pct() {
+	public Double getCommission_pct() {
 		return commission_pct;
 	}
 
-	public void setCommission_pct(Integer commission_pct) {
+	public void setCommission_pct(Double commission_pct) {
 		this.commission_pct = commission_pct;
 	}
 
@@ -126,5 +126,8 @@ public class Employees {
 	public String toString() {
 		return employee_id + " " + first_name + " " + last_name;
 	}
+
+
+	
 
 }
